Move directory of TestEntityGenerator
